requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.6.1
email-validator==2.2.0
Jinja2==3.1.4
reportlab==4.2.5
python-multipart==0.0.12
aiosmtplib==3.0.1
httpx==0.27.2
# Optional for Gemini content generation (fallback to no-op if missing)
google-generativeai==0.8.0

